Split optional LLM backend loading
