+++ todoist/web/api_components/automation_runtime.py
@@ -3,12 +3,22 @@
 
 # pylint: disable=protected-access,cyclic-import,too-many-lines,undefined-variable,global-variable-undefined,used-before-assignment,line-too-long,global-statement
 
-from __future__ import annotations
 
 from dataclasses import dataclass
+from collections.abc import Mapping, Sequence
+from datetime import datetime
+from http.server import BaseHTTPRequestHandler, ThreadingHTTPServer
+from pathlib import Path
 from typing import Any
 import contextlib
 import threading
+
+from google.oauth2.credentials import Credentials
+from omegaconf import DictConfig
+
+from todoist.automations.base import Automation
+from todoist.automations.observer import AutomationObserver
+from todoist.database.base import Database
 
 
 def _sync_api_globals() -> None:
